Prepare for SARIF integration, fix spelling, FF and Infer in ep_cpp.sh
Start to move away from line output parsing so there's a common format
 being used as much as possible for gathering static analysis issues
 data/results.
I chose SARIF since CPPCheck, FlawFinder, and Infer all support this
 format, and it is a format supported by a number of other static
 analysis tools that could then be integrated without consideration of
 any particular custom stdout format.
This opens up the possibility for other languages and tools that also
 support this format to be easily integrated into this Action.
The first step was ensuring all tools that can support it are using
 versions with SARIF support; this is complete, only cppcheck needed to
 be upgraded in the docker image that supports the runner, and is now
 set to version 2.17.1.
The second step in this was to install the Python 3 SARIF Object Model
 library into the docker image that supports the runner; this is
 complete.
The third step was to add the proper command line arguments to the
 tools that get called in the entrypoint_cpp.sh; this is also complete.
The fourth step was to add execution of a new python script that joins
 multiple SARIF files (one for each source file analyzed with cppcheck
 and one for each target directory analyzed with flawfinder) to
 entrypoint_cpp.sh; this is now complete also.
The fifth step is to actually write the new join_sarif Python 3 module,
 which is almost complete but not yet ready.
The final step is to integrate SARIF into the Action code and tests;
 this is not yet complete, but I am working on it and have already made
 significant progress, and should be something I can finish in the next
 2-3 weeks, time permitting.

Fixed spelling in the Action definition.

I finished integrating FlawFinder and Infer into entrypoint_cpp.sh
 script, which included adding -ff and -fi arguments to the call to the
 static_analysis_cpp Python 3 module.

Fixed/improved aspects of entrypoint_cpp.sh.
 - debug_print for get_files_to_check.py that says it is going to run
    the script now prints prior to it running so what is being printed
    aligns with what is actually happening during the script run.
 - Fixed an issue where all the header files in the project being
    analyzed were getting passed to cppcheck, which does not support
    direct analysis of header files.
 - Fix exclude_arg not working with cppcheck by removing it; it is
    unnecessary based on how I want cppcheck to be run.

Removed excess whitespace in test/test_utils.py.
Added trailing newline to test/utils/dummy_project/DummyFile.hpp.